Changed ShardSyncer to a public interface. Renamed implementation to KinesisShardSyncer.
